Structure GEDCOM lines while structuring document
Throughout the codebase, the FMP property builders require access to
the GEDCOM tags or XrefIds. This in turn requires helper functions
scattered throughout to parse the GEDCOM properties from the lines.

Since a queryable version of the GEDCOM is being constructed anyway, it
makes sense to structure the lines to be queryable at the same time.